Improve maintenance workflows
