--- ruff.toml
@@ -1,30 +1,0 @@
-line-length = 100
-target-version = "py314"
-
-[lint]
-select = [
-  "E",
-  "F",
-  "I",
-  "N",
-  "W",
-  "UP",
-  "S",
-  "B",
-  "A",
-  "C4",
-  "PT",
-  "RET",
-  "SIM",
-  "ERA",
-]
-ignore = []
-
-[lint.per-file-ignores]
-"test/*.py" = ["S101", "PT017", "S603", "S110", "N806"]
-"python/tests/*.py" = ["S101", "PT017", "N806"]
-"*.pyi" = ["A002"]
-
-[format]
-quote-style = "double"
-indent-style = "space"
